Fixing re-connector logic which spins off a new thread on failed initial connection every time Moving the functionality tests into a separate file Fixing the backend validation errors Adding more high-level coverage tests
