+++ tst/WebRTCClientTestFixture.h
@@ -6,10 +6,10 @@
 #define TEST_DEFAULT_REGION                                     ((PCHAR) "us-west-2")
 #define TEST_STREAMING_TOKEN_DURATION                           (40 * HUNDREDS_OF_NANOS_IN_A_SECOND)
 #define TEST_JITTER_BUFFER_CLOCK_RATE                           (1000)
-#define TEST_SIGNALING_MASTER_CLIENT_ID                         (PCHAR) "Test Master ClientId"
-#define TEST_SIGNALING_VIEWER_CLIENT_ID                         (PCHAR) "Test Viewer ClientId"
+#define TEST_SIGNALING_MASTER_CLIENT_ID                         (PCHAR) "Test_Master_ClientId"
+#define TEST_SIGNALING_VIEWER_CLIENT_ID                         (PCHAR) "Test_Viewer_ClientId"
 #define TEST_SIGNALING_CHANNEL_NAME                             (PCHAR) "ScaryTestChannel"
-#define SIGNAING_TEST_CORRELATION_ID                            (PCHAR) "Test correlation id"
+#define SIGNAING_TEST_CORRELATION_ID                            (PCHAR) "Test_correlation_id"
 
 //
 // Set the allocators to the instrumented equivalents
